Rename file training-hook.tsx to training-status-hook.tsx
