+++ src/App.tsx
@@ -19,7 +19,7 @@ import { stepsConfig } from "./steps-config";
 import { LoggingProvider } from "./logging/logging-hooks";
 import { ConnectionFlowProvider } from "./connections";
 import { GesturesProvider } from "./gestures-hooks";
-import { TrainingStatusProvider } from "./training-hook";
+import { TrainingStatusProvider } from "./training-status-hook";
 
 export interface ProviderLayoutProps {
   children: ReactNode;
